+++ flowmachine/flowmachine/core/init.py
@@ -152,7 +152,7 @@ def connect(
             host=redis_host, port=redis_port, password=redis_password
         )
         _start_threadpool(thread_pool_size=flowdb_connection_pool_size)
-        conn.available_dates(table="all", strictness=1, schema="events")
+        conn.available_dates
 
         print(f"FlowMachine version: {flowmachine.__version__}")