+++ flowmachine/flowmachine/core/server/server.py
@@ -33,51 +33,6 @@ logger = structlog.get_logger("flowmachine.debug", submodule=__name__)
 query_run_log = structlog.get_logger("flowmachine.query_run_log")
 
 
-async def update_available_dates(
-    *,
-    flowdb_connection: Connection,
-    pool: Executor,
-    sleep_time: int = 1800,
-    loop: bool = True,
-) -> None:
-    """
-    Background task to periodically refresh the cache entries for available dates
-    based on the content of the database.
-
-    Parameters
-    ----------
-    flowdb_connection : Connection
-        Flowdb connection to check dates on
-    pool : Executor
-        Executor to run the date check with
-    sleep_time : int, default 1800
-        Number of seconds to sleep for between checks
-    loop : bool, default True
-        Set to false to return after the first check
-
-    Returns
-    -------
-    None
-
-    """
-    check_func = partial(
-        flowdb_connection.available_dates.__wrapped__,
-        flowdb_connection,
-        table="all",
-        strictness=1,
-        schema="events",
-    )
-    while True:
-        logger.debug("Checking available dates.")
-        avail = await asyncio.get_running_loop().run_in_executor(
-            pool, check_func
-        )  # Note we're calling the uncached version here.
-        if not loop:
-            break
-        await asyncio.sleep(sleep_time)
-    return avail
-
-
 async def get_reply_for_message(
     *, msg_str: str, config: "FlowmachineServerConfig"
 ) -> ZMQReply:
@@ -287,11 +242,6 @@ async def recv(
     main_loop = asyncio.get_event_loop()
     main_loop.add_signal_handler(signal.SIGTERM, partial(shutdown, socket=socket))
 
-    main_loop.create_task(
-        update_available_dates(
-            flowdb_connection=flowdb_connection, pool=Query.thread_pool_executor
-        )
-    )
     main_loop.create_task(
         watch_and_shrink_cache(
             flowdb_connection=flowdb_connection,